fix: tried fixing the reload user function
